requirements.txt
pandas
numpy
scikit-learn
joblib
flask
torch
torchvision
Pillow